@@ -1,3 +1,3 @@
 [tool.black]
-target-version = ["py37"]
+target-version = ["py38"]
 line-length = 125
