@@ -27,8 +27,8 @@ from karapace.offset_watcher import OffsetWatcher
 from karapace.schema_models import ParsedTypedSchema, SchemaType, SchemaVersion, TypedSchema, ValidatedTypedSchema
 from karapace.schema_reader import KafkaSchemaReader
 from karapace.schema_references import Reference
-from karapace.typing import JsonData, ResolvedVersion, Subject, Version
-from typing import cast, Dict, List, Optional, Tuple, Union
+from karapace.typing import JsonObject, ResolvedVersion, Subject, Version
+from typing import Dict, List, Optional, Tuple, Union
 
 import asyncio
 import logging
@@ -129,7 +129,7 @@ class KarapaceSchemaRegistry:
         compatibility = self.database.get_subject_compatibility(subject=subject)
         if compatibility is None:
             # If no subject compatiblity found, use global compatibility
-            compatibility = cast(str, self.config["compatibility"])
+            compatibility = self.config["compatibility"]
         try:
             compatibility_mode = CompatibilityModes(compatibility)
         except ValueError as e:
@@ -259,7 +259,7 @@ class KarapaceSchemaRegistry:
             raise SchemasNotFoundException
         return schemas
 
-    def subject_version_get(self, subject: Subject, version: Version, *, include_deleted: bool = False) -> JsonData:
+    def subject_version_get(self, subject: Subject, version: Version, *, include_deleted: bool = False) -> JsonObject:
         validate_version(version)
         schema_versions = self.subject_get(subject, include_deleted=include_deleted)
         if not schema_versions:
@@ -272,7 +272,7 @@ class KarapaceSchemaRegistry:
         schema_id = schema_data.schema_id
         schema = schema_data.schema
 
-        ret = {
+        ret: JsonObject = {
             "subject": subject,
             "version": resolved_version,
             "id": schema_id,
